pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hci-extractor"
version = "0.1.0"
description = "Extract structured content from HCI academic papers"
authors = [{name = "HCI Research Team"}]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "google-generativeai>=0.3.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
hci-extractor = "hci_extractor.main:cli"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
strict = true

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]